Добавил проверки при регистрации, провел рефакторинг кода. Пожалуйста,
проверьте очень внимательно, мог случайно что-то заруинить

+++ src/main/java/com/example/StatMoney/entity/Asset.java
@@ -20,5 +20,5 @@ public class Asset {
     private String category;
     private double purchasePriceRub;             // Цена покупки в рублях
     private double purchasePriceUsd;             // Цена покупки в долларах
-    private int quantity;                    // Количество
+    private double quantity;                    // Количество
 }
